+++ adapter/inbound/packet.go
