+++ adapter/inbound/util.go
