+++ adapter/outbound/util.go
@@ -3,9 +3,7 @@ package outbound
 import (
 	"bytes"
 	"crypto/tls"
-	"fmt"
 	"net"
-	"net/url"
 	"strconv"
 	"sync"
 	"time"
@@ -15,42 +13,10 @@ import (
 	"github.com/Dreamacro/clash/transport/socks5"
 )
 
-const (
-	tcpTimeout = 5 * time.Second
-)
-
 var (
 	globalClientSessionCache tls.ClientSessionCache
 	once                     sync.Once
 )
-
-func urlToMetadata(rawURL string) (addr C.Metadata, err error) {
-	u, err := url.Parse(rawURL)
-	if err != nil {
-		return
-	}
-
-	port := u.Port()
-	if port == "" {
-		switch u.Scheme {
-		case "https":
-			port = "443"
-		case "http":
-			port = "80"
-		default:
-			err = fmt.Errorf("%s scheme not Support", rawURL)
-			return
-		}
-	}
-
-	addr = C.Metadata{
-		AddrType: C.AtypDomainName,
-		Host:     u.Hostname(),
-		DstIP:    nil,
-		DstPort:  port,
-	}
-	return
-}
 
 func tcpKeepAlive(c net.Conn) {
 	if tcp, ok := c.(*net.TCPConn); ok {
