+++ adapter/outboundgroup/util.go
