+++ rule/base.go
