+++ rule/domain.go
