+++ rule/domain_suffix.go
